% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_prompts.R
\name{context_history}
\alias{context_history}
\title{Generate a String Containing the Recent R Command History}
\usage{
context_history(
  prefix = "\\n## For Context the most recent list of commands run are :\\n",
  suffix = "\\n## End history of commands run.",
  n_commands = 10
)
}
\arguments{
\item{prefix}{Character string to prepend to the command history.
Defaults to "\\n## For Context the most recent list of commands run are:\\n".}

\item{suffix}{Character string to append to the command history.
Defaults to '\\n## End history of commands run.'.}

\item{n_commands}{Integer specifying the number of recent commands to
retrieve. Defaults to 10.}
}
\value{
A character string containing the formatted recent R command history.
}
\description{
Creates a string containing the recent R command history.
}
\details{
This function saves the R command history to a temporary file, retrieves the
most recent commands, and formats them into a single string, prepended with a
specified prefix and appended with a suffix.

The function, extracts the `n_commands` most recent commands, and prepends each
  command with its index. The commands are then combined into a single
  string using `txt_single()` and wrapped with the specified `prefix` and
  `suffix`. This is useful for providing context about recent actions in the
  R environment.
}
\examples{
\dontrun{
# Get the recent command history
history_string <- context_history()

# Print the history string
cat(history_string)
}
}
