% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_prompts.R
\name{context_r_version}
\alias{context_r_version}
\title{Generate a Context String Containing R Version Information}
\usage{
context_r_version(
  prefix = "\\n## For Context this information about the R version :\\n",
  suffix = "\\n## End of context for R version.",
  ...
)
}
\arguments{
\item{prefix}{Character string to prepend to the R version information.
Defaults to "\\n## For Context this information about the R version :\\n".}

\item{suffix}{Character string to append to the R version information.
Defaults to '\\n## End of context for R version.'.}

\item{...}{Additional arguments passed to `named_lst_to_string()`.}
}
\value{
A character string containing the formatted R version information.
}
\description{
Creates a formatted string of R version details.
}
\details{
This function extracts and formats the R version information into a string
suitable for providing context in debugging or code generation scenarios.

This function retrieves the `R.version` list, converts it into a
  formatted string using `named_lst_to_string()`, and wraps it with the
  specified `prefix` and `suffix`. This provides a concise way to include
  information about the R environment in context prompts or debugging
  messages.
}
\examples{
\dontrun{
# Get the R version context string
r_version_string <- context_r_version()

# Print the R version string
cat(r_version_string)
}
}
