% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_function_details.R
\name{extract_function_name}
\alias{extract_function_name}
\title{Extract Function Name}
\usage{
extract_function_name(func_code)
}
\arguments{
\item{func_code}{Character string with function code}
}
\value{
Character string with function name
}
\description{
Extracts function name from function definition
}
\examples{
  extract_function_name(".myFunc = function() {}")
  extract_function_name("f <<- function(...) NULL")
  extract_function_name("foo <- function(x) x + 1")
}
