% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_libraries.R
\name{functions_in_memory}
\alias{functions_in_memory}
\title{List Names of Functions Currently in Memory}
\usage{
functions_in_memory(.env = .GlobalEnv)
}
\arguments{
\item{.env}{Environment to search for functions. Defaults to the global environment.}
}
\value{
character A character vector containing the names of all functions currently in memory.
}
\description{
Returns a character vector of function names in memory.
}
\details{
This function identifies and returns the names of all functions currently loaded in the R environment.


The function works by first listing all objects in the current environment using `base::ls()`.
It then iterates through these objects, checking if each one is a function using `base::is.function()`.
Finally, it returns the names of the objects that are identified as functions.
This function is useful for introspection and debugging, allowing you to see which functions are currently available for use.
It leverages `purrr` for concise iteration and filtering.
}
\examples{
\dontrun{
# List all functions in the current environment
function_names <- functions_in_memory()
print(function_names)

# Check if a specific function is in memory
if ("mean" \%in\% functions_in_memory()) {
  print("The 'mean' function is in memory.")
} else {
  print("The 'mean' function is not in memory.")
}
}
}
