% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_testthat.R
\name{generate_testthat}
\alias{generate_testthat}
\title{Generate Testthat File for a Function}
\usage{
generate_testthat()
}
\value{
No return value. This function is called for its side effects.
}
\description{
This function automates the creation of a testthat file for a specified function.

Generates a testthat file by extracting function details, creating a test file, and appending test code.
}
\details{
The function uses `function_code_2` to extract details of the function under the cursor in the active RStudio document. It then calls `generate_testthat_from_function_details` to generate the test code.  The generated test code is saved in a file named `test-<function_name>.R` within the `tests/testthat` directory.  If the directory does not exist, it will be created. Finally, the function optionally opens the created test file in RStudio.
}
\examples{
\dontrun{
# Assuming you have a function defined in your script, place the cursor
# on the line containing the function definition and run this function.
generate_testthat()
}
}
