% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_libraries.R
\name{libraries_extract_file}
\alias{libraries_extract_file}
\title{Extract Libraries from a File}
\usage{
libraries_extract_file(file_name)
}
\arguments{
\item{file_name}{(character) The name of the file to extract library names from. This should be a character string representing the path to the file.}
}
\value{
(character) A character vector containing the names of the libraries extracted from the file.
}
\description{
Extracts R library names from a file.
}
\details{
This function extracts R library names from a specified file.

The function reads the entire content of the specified file, collapses it into a single string, and then uses `libraries_extract_parse()` to extract the library names. This is useful for identifying the dependencies of an R script or project by analyzing its source code.  It assumes that the file contains R code and that library calls are made using `library()` or `require()`.
}
\examples{
\dontrun{
# Create a dummy R file
cat("library(dplyr)\nrequire(ggplot2)\n", file = "temp.R")

# Extract libraries from the file
libraries <- libraries_extract_file("temp.R")
print(libraries)

# Clean up the dummy file
file.remove("temp.R")
}
}
