% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_libraries.R
\name{libraries_installed}
\alias{libraries_installed}
\title{List Installed R Packages}
\usage{
libraries_installed()
}
\value{
character A character vector containing the names of all installed R packages.
}
\description{
Lists all installed R packages.
}
\details{
This function retrieves a list of all installed R packages.

This function uses the `library()` function to get a list of installed packages and extracts the package names from the results. It returns a character vector containing the names of the installed packages. This function is useful for determining which packages are available for use in your R environment.
}
\examples{
\dontrun{
# Get a list of installed packages
installed_packages <- libraries_installed()

# Print the list of installed packages
print(installed_packages)

# Check if a specific package is installed
if ("dplyr" \%in\% installed_packages) {
  print("dplyr is installed")
} else {
  print("dplyr is not installed")
}
}
}
