% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template.R
\name{create_vibe_project}
\alias{create_vibe_project}
\title{Create a Vibe Project}
\usage{
create_vibe_project(
  path,
  project_name = "MyVibeProject",
  include_readme = TRUE,
  ...
)
}
\arguments{
\item{path}{Path where the project should be created}

\item{project_name}{Name of the project}

\item{include_readme}{Whether to include README}

\item{...}{Additional parameters from the template form}

\item{project_goal}{Description of the project goal}

\item{vibe_level}{The vibe level selected}

\item{music_genre}{Preferred music genre}
}
\description{
This function is called when creating a new Vibe project from the RStudio template.
For now, it just creates a standard R project structure.
}
