% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function.R
\name{generate_function_from_specs}
\alias{generate_function_from_specs}
\title{generate_function_from_specs}
\usage{
generate_function_from_specs(specs, function_name = "")
}
\arguments{
\item{specs}{}

\item{function_name}{}
}
\value{
the string returned from the LLM
}
\description{
generate_function_from_specs
}
