% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roxygen_comment.R
\name{generate_roxygen_comment}
\alias{generate_roxygen_comment}
\title{Generate Roxygen Comment}
\usage{
generate_roxygen_comment()
}
\value{
Inserts Roxygen comment block above the function.
}
\description{
Adds a Roxygen skeleton for the function at the current cursor position.
}
