% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roxygen_comment.R
\name{generate_roxygen_comment_from_code}
\alias{generate_roxygen_comment_from_code}
\title{generate_roxygen_comment_from_code}
\usage{
generate_roxygen_comment_from_code(func_code, ...)
}
\arguments{
\item{func_code}{}

\item{...}{}
}
\value{
the string returned from the LLM
}
\description{
generate_roxygen_comment_from_code
}
