% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_models_all}
\alias{llm_models_all}
\title{List All Available services with language Models}
\usage{
llm_models_all()
}
\value{
character A character vector containing the names of all available language models.
}
\description{
Lists all language models available in the 'ellmer' package.
}
\details{
Retrieves a list of all available services models within the 'ellmer' package.

This function identifies services models by listing objects within the 'ellmer' package that match the pattern "models\_". It then removes the "models\_" prefix, replaces underscores with spaces, and converts the resulting strings to title case for improved readability. This provides a user-friendly listing of available models.
}
\examples{
\dontrun{
# List all available language models
llm_models_all()
}
}
